### 2019-09-06 #1714 @whedon
[PRE REVIEW]: BayesPostEst: An R Package to Generate Postestimation Quantities for Bayesian MCMC Estimation
**Submitting author:** @jkarreth (<a href="http://orcid.org/0000-0003-4586-7153">Johannes Karreth</a>)
**Repository:** <a href="https://github.com/ShanaScogin/BayesPostEst" target ="_blank">https://github.com/ShanaScogin/BayesPostEst</a>
**Version:** 0.1.0
