### 2019-09-06 f0e4ac0
initial unit tests for mcmcMargEff and data creation
