+++ DESCRIPTION
@@ -51,7 +51,8 @@ Imports:
     stats,
     texreg,
     runjags,
-    tidyr (>= 0.5.1)
+    tidyr (>= 0.5.1),
+    HDInterval
 Depends:
     R (>= 2.14.0)
 Encoding: UTF-8
