### 2019-09-06 6504a9c
more files for mcmcMargEff unit tests
