### 2019-09-10 d02fadb
replace mcmcRocPrc with v2; remove ancillary artifcats
Preparation for pull request:

- replace mcmcRocPrc with the faster, improved version
- update tests, vignette, etc.
- remove previous profilign and benchmarking artifcasts
