+++ R/mcmcRocPrc.R
@@ -3,229 +3,41 @@
 #' Generate ROC and Precision-Recall curves after fitting a Bayesian logit or 
 #' probit regression.
 #' 
-#' @param modelmatrix model matrix, including intercept (if the intercept is 
-#'   among the parameters estimated in the model). Create with 
-#'   `model.matrix(formula, data)`. Note: the order of columns in the model 
-#'   matrix must correspond to the order of columns in the matrix of posterior 
-#'   draws in the \code{mcmcout} argument. See the `mcmcout`` argument for 
-#'   more and Beger (2016) for background.
-#' @param mcmcout posterior distributions of all logit coefficients, in matrix 
-#'   form. This can be created from rstan, MCMCpack, R2jags, etc. and 
-#'   transformed into a matrix using the function [coda::as.mcmc()] from the 
-#'   coda package for `jags`` class objects, as.matrix() from base R for \code{mcmc}, \code{mcmc.list}, \code{stanreg}, and 
-#'\code{stanfit} class objects, and \code{object$sims.matrix} for \code{bugs} class objects.
-#'Note: the order of columns in this matrix must correspond to the order of columns 
-#'in the model matrix. One can do this by examining the posterior distribution matrix and sorting the 
-#'variables in the order of this matrix when creating the model matrix. A useful function for sorting 
-#'column names containing both characters and numbers as 
-#'you create the matrix of posterior distributions is \code{mixedsort()} fom the gtools package.
-#' @param modelframe model frame in matrix form. Can be created using 
-#'   `as.matrix(model.frame(formula, data))`.
-#' @param curves logical indicator of whether or not to return values to plot the ROC or Precision-Recall 
-#'curves. If set to \code{FALSE} (default), results are returned as a list without the extra
-#'values. 
-#' @param link type of generalized linear model; a character vector set to 
-#'   `"logit"` (default) or `"probit"`.
+#' @param object A "rjags" object (see [R2jags::jags()]) for a fitted binary
+#'   choice model.
+#' @param yname (`character(1)`)\cr
+#'   The name of the dependent variable, should match the variable name in the 
+#'   JAGS data object.
+#' @param xnames ([base::character()])\cr
+#'   A character vector of the independent variable names, should match the 
+#'   corresponding names in the JAGS data object.
+#' @param curves logical indicator of whether or not to return values to plot 
+#'   the ROC or Precision-Recall curves. If set to \code{FALSE} (default), 
+#'   results are returned as a list without the extra values. 
 #' @param fullsims logical indicator of whether full object (based on all MCMC
 #'   draws rather than their average) will be returned. Default is `FALSE`. 
-#'   Note:  If `TRUE` is chosen, the function takes notably longer to execute.
+#'   Note: If `TRUE` is chosen, the function takes notably longer to execute.
 #' 
 #' @references Beger, Andreas. 2016. “Precision-Recall Curves.” Available at 
 #'   SSRN: [http://dx.doi.org/10.2139/ssrn.2765419](http://dx.doi.org/10.2139/ssrn.2765419)
 #' 
-#' @return This function returns a list with 4 elements:
-#'  - area_under_roc: area under ROC curve (scalar)
-#'  - area_under_prc: area under precision-recall curve (scalar)
-#'  - prc_dat: data to plot precision-recall curve (data frame)
-#'  - roc_dat: data to plot ROC curve (data frame)
+#' @return Returns a list; the specific structure depends on the combination
[…]
