+++ DESCRIPTION
@@ -25,7 +25,6 @@ Imports:
     R2jags,
     reshape2,
     rlang,
-    ROCR,
     stats,
     texreg,
     runjags,

[…]
