### 2019-09-11 #1722 @terrytangyuan
@openjournals/joss-eics Any idea on how to fix the compilation error?
