### 2019-09-18 8c5752f
setup initial pkgdown config
