### 2019-09-18 3f7ae6b
try out travis gh-pages build without PAT
