### 2019-09-20 #1722 @svmiller
I'll caution here that I am kind of clumsy on Github, and this is the first review of its kind I've done, but I believe I've completed my review for this paper. I obviously recommend publication (which is the language I'm used to using for social science papers) and I should've checked all boxes in my review toward that end.
