### 2019-09-26 #1722 @TiagoVentura
I checked all the boxes as well. The paper and the package are great, strongly recommend the publication. I just found on issue, that I append to this code. The function 'mcmcRocPrc' does not work as suggested in the Readme. I am providing an example here:
`rm(list=ls(all=TRUE))
library("remotes")
install_github("ShanaScogin/BayesPostEst")

df <- carData::Cowles

df$female <- (as.numeric(df$sex) - 2) * (-1)
df$volunteer <- as.numeric(df$volunteer) - 1
df$extraversion <- (df$extraversion - mean(df$extraversion)) / (2 * sd(df$extraversion))
df$neuroticism <- (df$neuroticism - mean(df$neuroticism)) / (2 * sd(df$neuroticism))

dl <- as.list(df[, c("volunteer", "female", "neuroticism", "extraversion")])
dl$N <- nrow(df)

mod.jags <- paste("
model {
for (i in 1:N){
  volunteer[i] ~ dbern(p[i])
  logit(p[i]) <- mu[i]
  mu[i] <- b[1] + b[2] * female[i] + b[3] * neuroticism[i] + b[4] * extraversion[i]
  }

for(j in 1:4){
  b[j] ~ dnorm(0, 0.1)
  }

}
")
writeLines(mod.jags, "mod.jags")

params.jags <- c("b")
inits1.jags <- list("b" = rep(0, 4))
inits.jags <- list(inits1.jags, inits1.jags, inits1.jags, inits1.jags)

library("R2jags")
set.seed(123)

fit.jags <- jags(data = dl, inits = inits.jags,
                 parameters.to.save = params.jags, n.chains = 4, n.iter = 2000,
                 n.burnin = 1000, model.file = "mod.jags")

mcmcmat.jags <- as.matrix(coda::as.mcmc(fit.jags))

mf <- model.frame(volunteer ~ female + neuroticism + extraversion, data = df)

fitstats <- mcmcRocPrc(modelmatrix = mm,
                       modelframe = mf,
                       mcmcout = mcmcmat.jags[, 1:ncol(mm)],
                       curves = TRUE,
                       link = "logit",
                       fullsims = FALSE)`

The error:

> Error in mcmcRocPrc(modelmatrix = mm, modelframe = mf, mcmcout = mcmcmat.jags[,  :
  unused arguments (modelmatrix = mm, modelframe = mf, mcmcout = mcmcmat.jags[, 1:ncol(mm)], link = "logit")

Let me know if it is something in my end.
