### 2019-09-26 #1722 @terrytangyuan
@jkarreth Could you take a look at the above issue pointed out?
