### 2019-09-26 #1722 @jkarreth
@TiagoVentura Thank you for pointing this out! The function is ahead of the Readme and vignette here, so I need to adjust the code in these documents. I'll update both documents and will post here again when that is done.
