### 2019-09-26 #1722 @TiagoVentura
Yes! That was my intuition of what was happening @jkarreth . Thanks!
