### 2019-09-26 769e42c
updated mcmcRocPrc example to match new function
