### 2019-09-26 c8f2129
updated mcmcRocPrc example to match new function
