+++ vignettes/references.bib
@@ -6,6 +6,22 @@
 
 %% Saved with string encoding Unicode (UTF-8) 
 
+@Manual{rjags,
+    title = {rjags: Bayesian Graphical Models using MCMC},
+    author = {Martyn Plummer},
+    year = {2018},
+    note = {R package version 4-8},
+    url = {https://CRAN.R-project.org/package=rjags},
+  }
+
+@Manual{R2jags,
+    title = {R2jags: Using R to Run 'JAGS'},
+    author = {Yu-Sung Su and Masanao Yajima,},
+    year = {2015},
+    note = {R package version 0.5-7},
+    url = {https://CRAN.R-project.org/package=R2jags},
+  }
+
 @article{Kruschke2018,
 	author = {John K. Kruschke},
 	title ={Rejecting or Accepting Parameter Values in Bayesian Estimation},
