+++ vignettes/getting_started.Rmd
@@ -4,13 +4,16 @@ author:
   - "Johannes Karreth"
   - "Shana Scogin"
   - "Rob Williams"
+  - "Andreas Beger"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 bibliography: "references.bib"
 vignette: >
   %\VignetteIndexEntry{getting_started}
-  %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+  %\VignetteEngine{knitr::rmarkdown}
+editor_options: 
+  chunk_output_type: console
 ---
 
 
@@ -51,7 +54,7 @@ Most functions in this package work with posterior distributions of parameters. 
 
 # Example data
 
-This vignette uses the `Cowles` dataset (@CowlesDavis1987) from the carData package (@carData2018). 
+This vignette uses the `Cowles` dataset [@CowlesDavis1987] from the carData package [@carData2018]. 
 
 ```{r}
 df <- carData::Cowles
@@ -87,7 +90,7 @@ To use `BayesPostEst`, we first estimate a Bayesian regression model. This vigne
 
 ## JAGS
 
-First, we prepare the data for JAGS (@jags2017). Users need to combine all variables into a list and specify any other elements, like in this case N, the number of observations.
+First, we prepare the data for JAGS [@jags2017]. Users need to combine all variables into a list and specify any other elements, like in this case N, the number of observations.
 
 ```{r}
 dl <- as.list(df[, c("volunteer", "female", "neuroticism", "extraversion")])
@@ -120,7 +123,7 @@ inits1.jags <- list("b" = rep(0, 4))
 inits.jags <- list(inits1.jags, inits1.jags, inits1.jags, inits1.jags)
 ```
 
-Now, fit the model.
+Now, fit the model using the R2jags [@R2jags] package. <!-- or rjags [@rjags] packages. -->
 
 ```{r}
 library("R2jags")
@@ -130,9 +133,18 @@ fit.jags <- jags(data = dl, inits = inits.jags,
   n.burnin = 1000, model.file = "mod.jags")
 ```
 
+<!-- ```{r} -->
+<!-- library("rjags") -->
+<!-- mod.rjags <- jags.model(file = "mod.jags", data = dl, inits = inits.jags,  -->
+<!--                         n.chains = 4, n.adapt = 1000) -->
+<!-- fit.rjags <- coda.samples(model = mod.rjags,  -->
+<!--                           variable.names = params.jags,  -->
+<!--                           n.iter = 2000) -->
+<!-- ``` -->
+
 ## MCMCpack
 
-We estimate the same model using MCMCpack (@MCMCpack).
+We estimate the same model using MCMCpack [@MCMCpack].
 
 ```{r}
 library("MCMCpack")
@@ -183,7 +195,7 @@ fit.stan <- stan(file = "mod.stan",
            pars = c("b"),     
            chains = 4,        
            iter = 2000,       
-           seed = 123)        
+           seed = 123)    
 ```
 
 ## rstanarm
@@ -220,6 +232,10 @@ mcmcTab(fit.jags)
 ```
 
 ```{r}
+mcmcTab(fit.rjags)
+```
+
+```{r}
 mcmcTab(fit.MCMCpack)
 ```
 
@@ -241,7 +257,7 @@ mcmcTab(fit.jags, Pr = TRUE)
 
 ### ROPE
 
[…]
