### 2019-09-27 5d670e7
uncommented rjags model
