+++ vignettes/getting_started.Rmd
@@ -86,7 +86,7 @@ $$
 
 # Model estimation
 
-To use `BayesPostEst`, we first estimate a Bayesian regression model. This vignette demonstrates four tools for doing so: JAGS (via the R2jags package), MCMCpack, and the two Stan interfaces rstan and rstanarm.
+To use `BayesPostEst`, we first estimate a Bayesian regression model. This vignette demonstrates five tools for doing so: JAGS (via the [R2jags](https://cran.r-project.org/package=R2jags) and [rjags](https://cran.r-project.org/package=rjags) packages), [MCMCpack](https://cran.r-project.org/package=MCMCpack), and the two Stan interfaces [rstan](https://cran.r-project.org/package=rstan) and [rstanarm](https://cran.r-project.org/package=rstanarm).
 
 ## JAGS
 
@@ -123,7 +123,7 @@ inits1.jags <- list("b" = rep(0, 4))
 inits.jags <- list(inits1.jags, inits1.jags, inits1.jags, inits1.jags)
 ```
 
-Now, fit the model using the R2jags [@R2jags] package. <!-- or rjags [@rjags] packages. -->
+Now, fit the model using the R2jags package. 
 
 ```{r}
 library("R2jags")
@@ -133,14 +133,16 @@ fit.jags <- jags(data = dl, inits = inits.jags,
   n.burnin = 1000, model.file = "mod.jags")
 ```
 
-<!-- ```{r} -->
-<!-- library("rjags") -->
-<!-- mod.rjags <- jags.model(file = "mod.jags", data = dl, inits = inits.jags,  -->
-<!--                         n.chains = 4, n.adapt = 1000) -->
-<!-- fit.rjags <- coda.samples(model = mod.rjags,  -->
-<!--                           variable.names = params.jags,  -->
-<!--                           n.iter = 2000) -->
-<!-- ``` -->
+The same data and model can be used to fit the model using the rjags package:
+
+```{r}
+library("rjags")
+mod.rjags <- jags.model(file = "mod.jags", data = dl, inits = inits.jags,
+                        n.chains = 4, n.adapt = 1000)
+fit.rjags <- coda.samples(model = mod.rjags,
+                          variable.names = params.jags,
+                          n.iter = 2000)
+```
 
 ## MCMCpack
 
