### 2019-09-27 806b2b8
uncommented rjags model
