### 2019-09-27 6931ef1
added rjags to Suggests
