### 2019-09-27 7958cba
added author contributions
