### 2019-09-27 #1722 @jkarreth
@TiagoVentura Thank you again for finding this issue! I updated the Readme to reflect the new arguments to `mcmcRocPrc`. The vignette should already be correct. Would you mind taking a quick look at the new code and see if you can reproduce it? The relevant part is:

```
fitstats <- mcmcRocPrc(object = fit.jags,
                       yname  = "volunteer",
                       xnames = c("female", "neuroticism", "extraversion"),
                       curves = TRUE,
                       fullsims = FALSE)
```
