### 2019-09-27 #1722 @TiagoVentura
All good here. No more issues with the code. Thank you for the opportunity to review the paper, Looking forward to using the package!
