### 2019-09-27 #1722 @jkarreth
Wonderful! Thank you very much @TiagoVentura and @svmiller for your time and efforts.

@terrytangyuan, I just made a small update to the paper.md (added explicit description of author contributions and changed an author affiliation). Is there anything else I should be doing now?
