### 2019-09-28 2c61621
fixed typos, added bayestestR reference
