+++ JOSSpaper/paper.bib
@@ -6,6 +6,20 @@
 
 %% Saved with string encoding Unicode (UTF-8) 
 
+@Article{bayestestR,
+    title = {bayestestR: Describing Effects and their Uncertainty,
+      Existence and Significance within the Bayesian Framework.},
+    author = {Dominique Makowski and Mattan S. Ben-Shachar and Daniel
+      Lüdecke},
+    journal = {Journal of Open Source Software},
+    doi = {10.21105/joss.01541},
+    year = {2019},
+    number = {40},
+    volume = {4},
+    pages = {1541},
+    url = {https://joss.theoj.org/papers/10.21105/joss.01541},
+  }
+
 @Article{brms,
     title = {{brms}: An {R} Package for {Bayesian} Multilevel Models Using {Stan}},
     author = {Paul-Christian Bürkner},
