### 2019-09-28 ce0d643
resized table figure to 50%
