### 2019-09-28 #1722 @jkarreth
I just fixed a formatting issue in the paper and am recompiling - hope that's OK!
