### 2019-09-28 684e8b8
fixed math formatting
