+++ JOSSpaper/paper.md
@@ -114,7 +114,7 @@ $$
 \text{Pr}(y = 1 | x_{k}) = \Phi(\beta_1 + \beta_2 x_{k})
 $$
 
-To evaluate the relationship between covariates and a binary outcome, `mcmcAveProb` calculates the predicted probability ($Pr(y = 1)$) at pre-defined values of one covariate of interest ($x$), while all other covariates are held at a "typical" value. This follows suggestions outlined in @KingEtal2000 and elsewhere, which are commonly adopted by users of GLMs. The `mcmcAveProb` function by default calculates the median value of all covariates other than $x$ as "typical" values. Users can then access the full posterior distributions of the predicted probability at different values of $x$, or the median and a credible interval of the same quantity.
+To evaluate the relationship between covariates and a binary outcome, `mcmcAveProb` calculates the predicted probability ($\text{Pr}(y = 1)$) at pre-defined values of one covariate of interest ($x$), while all other covariates are held at a "typical" value. This follows suggestions outlined in @KingEtal2000 and elsewhere, which are commonly adopted by users of GLMs. The `mcmcAveProb` function by default calculates the median value of all covariates other than $x$ as "typical" values. Users can then access the full posterior distributions of the predicted probability at different values of $x$, or the median and a credible interval of the same quantity.
 
 As an alternative to probabilities for "typical" cases, @HanmerKalkan2013 suggest to calculate predicted probabilities for all observed cases and then derive an "average effect". In their words, the goal of this postestimation "is to obtain an estimate of the average effect in the population ... rather than seeking to understand the effect for the average case." BayesPostEst allows users to calculate these quantities for binary and continuous predictors based on observed values of covariates, using the `mcmcObsProb` function. Users can then summarize or visualize these quantities as is common in the social sciences [@Long1997; @KingEtal2000; @HanmerKalkan2013] 
 
