### 2019-09-28 #1722 @ShanaScogin
@terrytangyuan Here is the Zenodo doi associated with this release: http://doi.org/10.5281/zenodo.3464071

How's that look? Do you need anything else from me? Everything should be up to date with was was linked here. (Also, thank you again everyone for the time and feedback!!!)
