### 2019-09-30 #1722 @terrytangyuan
@ShanaScogin @jkarreth Could you fix the above missing DOI?
