### 2019-09-30 ff361a1
added doi for Clarify paper/replication archive
