+++ JOSSpaper/paper.bib
@@ -6,136 +6,20 @@
 
 %% Saved with string encoding Unicode (UTF-8) 
 
-@Article{bayestestR,
-    title = {bayestestR: Describing Effects and their Uncertainty,
-      Existence and Significance within the Bayesian Framework.},
-    author = {Dominique Makowski and Mattan S. Ben-Shachar and Daniel
-      Lüdecke},
-    journal = {Journal of Open Source Software},
-    doi = {10.21105/joss.01541},
-    year = {2019},
-    number = {40},
-    volume = {4},
-    pages = {1541},
-    url = {https://joss.theoj.org/papers/10.21105/joss.01541},
+@Manual{rjags,
+    title = {rjags: Bayesian Graphical Models using MCMC},
+    author = {Martyn Plummer},
+    year = {2018},
+    note = {R package version 4-8},
+    url = {https://CRAN.R-project.org/package=rjags},
   }
 
-@Article{brms,
-    title = {{brms}: An {R} Package for {Bayesian} Multilevel Models Using {Stan}},
-    author = {Paul-Christian Bürkner},
-    journal = {Journal of Statistical Software},
-    year = {2017},
-    volume = {80},
-    number = {1},
-    pages = {1--28},
-    doi = {10.18637/jss.v080.i01},
-  }
-
-@Article{ggmcmc,
-    title = {{ggmcmc}: Analysis of {MCMC} Samples and {B}ayesian Inference},
-    author = {Xavier Fern{\'a}ndez-i-Mar{\'i}n},
-    journal = {Journal of Statistical Software},
-    year = {2016},
-    volume = {70},
-    number = {9},
-    pages = {1--20},
-    doi = {10.18637/jss.v070.i09},
-  }
-
-@Manual{knitr,
-    title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
-    author = {Yihui Xie},
-    year = {2019},
-    note = {R package version 1.23},
-    url = {https://yihui.name/knitr/},
-  }
-
-@Manual{xtable,
-    title = {xtable: Export Tables to LaTeX or HTML},
-    author = {David B. Dahl and David Scott and Charles Roosen and Arni Magnusson and Jonathan Swinton},
-    year = {2019},
-    note = {R package version 1.8-4},
-    url = {https://CRAN.R-project.org/package=xtable},
-  }
-
-@Manual{flextable,
-    title = {flextable: Functions for Tabular Reporting},
-    author = {David Gohel},
-    year = {2019},
-    note = {R package version 0.5.5},
-    url = {https://CRAN.R-project.org/package=flextable},
-  }
-
-@article{Gelman2008,
-	author = {Gelman, Andrew},
-	title = {Scaling regression inputs by dividing by two standard deviations},
-	journal = {Statistics in Medicine},
-	volume = {27},
-	number = {15},
-	pages = {2865-2873},
-	doi = {10.1002/sim.3107},
-	year = {2008}
-}
-
-@Article{texreg,
-    title = {{texreg}: Conversion of Statistical Model Output in {R} to {\LaTeX} and {HTML} Tables},
-    author = {Philip Leifeld},
-    journal = {Journal of Statistical Software},
-    year = {2013},
-    volume = {55},
-    number = {8},
-    pages = {1--24},
-    url = {http://www.jstatsoft.org/v55/i08/},
-  }
-
-@Manual{sjPlot,
[…]
