### 2019-09-30 #1722 @jkarreth
@terrytangyuan done, this is now added to paper.bib
