### 2019-09-30 #1722 @jkarreth
I'm not sure why this keeps happening, but if it's OK with you, @terrytangyuan, that DOI is not essential. It's attached to the replication data for the article we cite (and the article itself doesn't have a DOI since it's from 2000), so the citation should still be correct even without DOI.
