### 2019-09-30 #1722 @kyleniemeyer
Whedon's DOI warnings are only meant to be warnings, so if you checked on one manually and it doesn't look like an actual issue, we can ignore
