### 2019-09-30 #1722 @terrytangyuan
Thanks @kyleniemeyer. Since you are already here, could you take it from here? The paper looks good to me.
