### 2019-10-01 #1722 @kyleniemeyer
Hi @jkarreth,

We're nearly done, just a few typos/fixes for the paper:
- in your author list, some of the affiliations are missing states
- in your code snippet on page 2, some of the commented text is cut off; I would suggest rearranging and moving stuff to new (shorter) lines
- on page 5, "or observed 1s" seems to be a typo (or is otherwise not clear); should "1s" be "ones"?
- add commas after "e.g" throughout the document: "e.g." -> "e.g.,"
- in the Beger reference, the URL given should be removed and converted to a DOI (stripping out the "http://dx.doi.org/")
- in multiple references, words in titles need `{}` to protect capitalization ("Monte Carlo", "R", etc.)
