+++ JOSSpaper/paper.bib
@@ -7,7 +7,7 @@
 %% Saved with string encoding Unicode (UTF-8) 
 
 @Manual{rjags,
-    title = {rjags: Bayesian Graphical Models using MCMC},
+    title = {rjags: Bayesian Graphical Models using {MCMC}},
     author = {Martyn Plummer},
     year = {2018},
     note = {R package version 4-8},
@@ -15,7 +15,7 @@
   }
 
 @Manual{R2jags,
-    title = {R2jags: Using R to Run 'JAGS'},
+    title = {R2jags: Using {R} to Run {JAGS}},
     author = {Yu-Sung Su and Masanao Yajima,},
     year = {2015},
     note = {R package version 0.5-7},
@@ -34,64 +34,49 @@
 
 @article{Beger2016,
 	Author = {Beger, Andreas},
-	Date-Added = {2019-07-19 07:30:03 -0400},
-	Date-Modified = {2019-07-19 07:30:54 -0400},
 	Journal = {Available at SSRN: https://ssrn.com/abstract=2765419},
 	Title = {Precision-Recall Curves},
-	Url = {http://dx.doi.org/10.2139/ssrn.2765419},
+	Doi = {10.2139/ssrn.2765419},
 	Year = {2016}}
 
 @article{CowlesDavis1987,
 	Author = {Cowles, Michael and Davis, Caroline},
-	Date-Added = {2019-07-18 22:37:30 -0400},
-	Date-Modified = {2019-07-18 22:37:30 -0400},
 	Doi = {10.1111/j.2044-8309.1987.tb00769.x},
 	Journal = {British Journal of Social Psychology},
 	Number = {2},
 	Pages = {97-102},
 	Title = {The subject matter of psychology: Volunteers},
 	Volume = {26},
-	Year = {1987},
-	Bdsk-Url-1 = {https://doi.org/10.1111/j.2044-8309.1987.tb00769.x}}
+	Year = {1987}}
 
 @article{HanmerKalkan2013,
 	Author = {Hanmer, Michael J. and Kalkan, Kerem Ozan},
-	Date-Added = {2014-06-03 21:16:35 +0000},
-	Date-Modified = {2018-07-11 00:38:06 +0000},
 	Doi = {10.1111/j.1540-5907.2012.00602.x},
 	Journal = {American Journal of Political Science},
 	Number = {1},
 	Pages = {263-277},
 	Title = {{Behind the Curve: Clarifying the Best Approach to Calculating Predicted Probabilities and Marginal Effects from Limited Dependent Variable Models}},
 	Volume = {57},
-	Year = {2013},
-	Bdsk-Url-1 = {http://dx.doi.org/10.1111/j.1540-5907.2012.00602.x}}
+	Year = {2013}}
 
 @article{jags2017,
 	Author = {Plummer, Martyn},
-	Date-Added = {2019-07-19 07:36:35 -0400},
-	Date-Modified = {2019-07-19 07:37:09 -0400},
-	Title = {{JAGS Version 4.3.0 User Manual}},
+	Title = {{JAGS} Version 4.3.0 User Manual},
 	Url = {http://mcmc-jags.sourceforge.net},
 	Year = {2017}}
 
 @article{Karreth2018,
 	Author = {Karreth,Johannes},
-	Date-Added = {2017-10-09 13:51:15 +0000},
-	Date-Modified = {2018-05-17 00:55:28 +0000},
 	Doi = {10.1080/03050629.2018.1389728},
 	Journal = {International Interactions},
 	Number = {3},
 	Pages = {463-490},
 	Title = {The Economic Leverage of International Organizations in Interstate Disputes},
[…]
