### 2019-10-01 #1722 @jkarreth
Thank you for the close read and for finding these issues, @kyleniemeyer ! These should now all be fixed in https://github.com/ShanaScogin/BayesPostEst/tree/master/JOSSpaper:

- [x] in your author list, some of the affiliations are missing states
- [x] in your code snippet on page 2, some of the commented text is cut off; I would suggest rearranging and moving stuff to new (shorter) lines
- [x] on page 5, "or observed 1s" seems to be a typo (or is otherwise not clear); should "1s" be "ones"?
- [x] add commas after "e.g" throughout the document: "e.g." -> "e.g.,"
- [x] in the Beger reference, the URL given should be removed and converted to a DOI (stripping out the "http://dx.doi.org/")
- [x] in multiple references, words in titles need {} to protect capitalization ("Monte Carlo", "R", etc.)
