### 2019-10-01 #1722 @jkarreth
Not sure how, but some of the references were lost in the latest .bib file. I just re-inserted them. Sorry about that!
